refactor(mitm-addon): separate usage flush summaries
